tests: Make JUnit tests compatible to the native JNI implementation.
Some of the JUnit tests were written explicitely for the Java implemenation
and are not compatible with the Native implementation. Those tests are
requesting a Java implementation from the ClientFactory now.
