+++ java/servers/src/org/xtreemfs/utils/xtfs_cleanup_osd.java
@@ -14,6 +14,7 @@ import java.util.Map;
 import org.xtreemfs.common.libxtreemfs.AdminClient;
 import org.xtreemfs.common.libxtreemfs.ClientFactory;
 import org.xtreemfs.common.libxtreemfs.Options;
+import org.xtreemfs.common.libxtreemfs.ClientFactory.ClientType;
 import org.xtreemfs.foundation.SSLOptions;
 import org.xtreemfs.foundation.logging.Logging;
 import org.xtreemfs.foundation.pbrpc.Schemes;
@@ -191,7 +192,7 @@ public class xtfs_cleanup_osd {
         Options userOptions = new Options();
         UserCredentials userCredentials = UserCredentials.newBuilder().setUsername("root").addGroups("root")
                 .build();
-        client = ClientFactory.createAdminClient(dirAddrs, userCredentials, sslOptions, userOptions);
+        client = ClientFactory.createAdminClient(ClientType.JAVA, dirAddrs, userCredentials, sslOptions, userOptions);
         try {
             client.start();
         } catch (Exception e) {
